// angular/layer-creation/src/modules/luciad/interfaces/UILayerTypes.ts
export enum UILayerTypes {
  WMSLayer = 'WMSLayer',
  PanoramicLayer = 'PanoramicLayer',
  HSPCLayer = 'HSPCLayer',
  OGC3DTILES = 'OGC3DTILES',
  LTSLayer = 'LTSLayer',
  TMSLayer = 'TMSLayer',
  WFSLayer = 'WFSLayer',
  BingmapsLayer = 'BingmapsLayer',
  LayerGroup = 'LayerGroup',
  KMLLayer = 'KMLLayer',
  SlicingBoxLayer = 'SlicingBoxLayer',
}